requirements.txt
# Core dependencies
fastapi>=0.115.0
uvicorn>=0.30.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Redis
redis>=5.0.0
fakeredis>=2.20.0  # For testing

# Database
sqlalchemy>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0  # For FastAPI TestClient
pytest-cov>=4.1.0  # For coverage reports

# API dependencies
websockets>=12.0  # For WebSocket support
python-multipart>=0.0.9  # For form data parsing

# Utilities
coloredlogs>=15.0.1
python-dateutil>=2.8.2

# Langroid for sentiment analysis and master agent
langroid>=0.40.0

# Data processing for MACD
numpy>=1.26.0
pandas>=2.2.0

# Monitoring
prometheus-client>=0.17.1

# Scheduling
apscheduler>=3.10.4

# Trading and data acquisition
ccxt>=4.0.0
youtube-transcript-api>=0.6.2
yt-dlp>=2023.0.0
openai-whisper>=20231117